requirements.txt
# Core dependencies for RL Maze Solver
# Numerical computing and array operations
numpy>=1.21.0,<2.1.0

# Data manipulation and analysis (for matrix display)
pandas>=1.3.0,<3.0.0

# Image processing and visualization
Pillow>=8.0.0,<12.0.0

# Optional: For development and testing
# pytest>=6.0.0  # Uncomment for running tests
# black>=21.0.0   # Uncomment for code formatting
# flake8>=3.8.0   # Uncomment for linting
